refactor: adopt .client.tsx convention for client components loaded with next/dynamic
