--- app/[locale]/_components/color-scheme-select-loading-indicator.tsx
@@ -1,5 +1,0 @@
-import type { ReactNode } from "react";
-
-export function ColorSchemeSelectLoadingIndicator(): ReactNode {
-	return <div className="size-10" />;
-}
